Added Simple Security for v1.3
Added simple security setting for Score Entry webpage.  What this
entails is to have the access to the webpage protected by a web service
(and have the web service return the contents of the web page if access
is granted).  User enters a password to gain access, and must match
password set in application.