Score_Entry: Fix issue with reset and previous data
The reset form function needs to be sure to remove the qwarn
class from the question divs. If not, a warning due to mismatch to
previously recorded answer will persist across submits.
Also added warn and old classes (were already in the html) to
highlight and unhighlight the previous value.